refactor(code-structure): remove unnecessary functions, fix extractSecondParam, major code structure refactor
