+++ ircserver/includes/Server.hpp
@@ -1,14 +1,6 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   Server.hpp                                         :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: dpetrukh <dpetrukh@student.42lisboa.com    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/06/27 10:58:19 by jqueijo-          #+#    #+#             */
-/*   Updated: 2025/08/29 12:08:11 by dpetrukh         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
+//
+//
+//
 
 #ifndef SERVER_HPP
 # define SERVER_HPP
@@ -20,32 +12,15 @@
 
 class Server {
 	public:
-		static const int BUFFER_SIZE = 1024;
-
-		// Constructor
 		Server();
 
-		// Lifecycle & Core Setup
 		void serverInit(const std::string& port, const std::string& password);
-		void serverSocketCreate();
 		static void signalHandler(int signum);
 		void closeFds();
-		void clearUser(int fd);
-
-		// Connection Handling
-		void acceptNewUser();
-		void receiveNewData(int fd);
-
-		void handleRawMessage(int fd, const char* buffer);
-
-		// Command Handlers
-		void handlePassCommand(User &user, std::string cmdParameters);
-		void handleNickCommand(User &user, std::string cmdParameters);
-		void handleUserCommand(User &user, std::string cmdParameters);
-		void handleJoinCommand(User &user, const std::string& commandParams);
-		void handleTopicCommand(User &user, const std::string& commandParams);
 
 	private:
+		static const int BUFFER_SIZE = 1024;
+
 		// Server State
 		std::string name;
 		int serverSocketFd;
@@ -59,6 +34,22 @@ class Server {
 		std::list<User> users;
 		std::vector<struct pollfd> pollFds;
 		std::list<Channel> channels;
+
+		// Lifecycle & Core Setup
+		void serverSocketCreate();
+		void clearUserFromPoll(int fd);
+
+		// Connection Handling
+		void acceptNewUser();
+		void receiveNewData(int fd);
+		void handleRawMessage(int fd, const char* buffer);
+
+		// Command Handlers
+		void handlePassCommand(User &user, std::string cmdParameters);
+		void handleNickCommand(User &user, std::string cmdParameters);
+		void handleUserCommand(User &user, std::string cmdParameters);
+		void handleJoinCommand(User &user, const std::string& commandParams);
+		void handleTopicCommand(User &user, const std::string& commandParams);
 
 		// Channel Utilities
 		Channel& getChannel(User& targetUser, const std::string& channelName);
