+++ ircserver/srcs/main.cpp
@@ -1,14 +1,6 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   main.cpp                                           :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: jqueijo- <jqueijo-@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/06/27 12:21:39 by jqueijo-          #+#    #+#             */
-/*   Updated: 2025/07/09 09:43:42 by jqueijo-         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
+//
+//
+//
 
 #include "../includes/User.hpp"
 #include "../includes/Server.hpp"
